+++ libs/putty/sshpubk.c
@@ -38,7 +38,7 @@ static int loadrsakey_main(FILE * fp, struct RSAKey *key, int pub_only,
     *error = NULL;
 
     /* Slurp the whole file (minus the header) into a buffer. */
-    len = fread(buf, 1, sizeof(buf), fp);
+    len = (int)fread(buf, 1, sizeof(buf), fp);
     fclose(fp);
     if (len < 0 || len == sizeof(buf)) {
 	*error = "error reading file";
@@ -108,7 +108,7 @@ static int loadrsakey_main(FILE * fp, struct RSAKey *key, int pub_only,
      */
     if (ciphertype) {
 	MD5Init(&md5c);
-	MD5Update(&md5c, (unsigned char *)passphrase, strlen(passphrase));
+	MD5Update(&md5c, (unsigned char *)passphrase, (unsigned int)strlen(passphrase));
 	MD5Final(keybuf, &md5c);
 	des3_decrypt_pubkey(keybuf, buf + i, (len - i + 7) & ~7);
 	smemclr(keybuf, sizeof(keybuf));	/* burn the evidence */
@@ -410,9 +410,9 @@ int saversakey(const Filename *filename, struct RSAKey *key, char *passphrase)
      */
     if (passphrase) {
 	MD5Init(&md5c);
-	MD5Update(&md5c, (unsigned char *)passphrase, strlen(passphrase));
+	MD5Update(&md5c, (unsigned char *)passphrase, (unsigned int)strlen(passphrase));
 	MD5Final(keybuf, &md5c);
-	des3_encrypt_pubkey(keybuf, estart, p - estart);
+	des3_encrypt_pubkey(keybuf, estart, (int)(p - estart));
 	smemclr(keybuf, sizeof(keybuf));	/* burn the evidence */
     }
 
@@ -583,7 +583,7 @@ static unsigned char *read_blob(FILE * fp, int nlines, int *bloblen)
 	    sfree(blob);
 	    return NULL;
 	}
-	linelen = strlen(line);
+	linelen = (int)strlen(line);
 	if (linelen % 4 != 0 || linelen > 64) {
 	    sfree(blob);
 	    sfree(line);
@@ -631,7 +631,7 @@ const struct ssh_signkey *find_pubkey_alg_len(int namelen, const char *name)
 
 const struct ssh_signkey *find_pubkey_alg(const char *name)
 {
-    return find_pubkey_alg_len(strlen(name), name);
+		return find_pubkey_alg_len((int)strlen(name), name);
 }
 
 struct ssh2_userkey *ssh2_load_userkey(const Filename *filename,
@@ -646,7 +646,7 @@ struct ssh2_userkey *ssh2_load_userkey(const Filename *filename,
     unsigned char *public_blob, *private_blob;
     int public_blob_len, private_blob_len;
     int i, is_mac, old_fmt;
-    int passlen = passphrase ? strlen(passphrase) : 0;
+    int passlen = passphrase ? (int)strlen(passphrase) : 0;
     const char *error = NULL;
 
     ret = NULL;			       /* return NULL for most errors */
@@ -786,9 +786,9 @@ struct ssh2_userkey *ssh2_load_userkey(const Filename *filename,
 	    free_macdata = 0;
 	} else {
 	    unsigned char *p;
-	    int namelen = strlen(alg->name);
-	    int enclen = strlen(encryption);
-	    int commlen = strlen(comment);
+			int namelen = (int)strlen(alg->name);
+			int enclen = (int)strlen(encryption);
+			int commlen = (int)strlen(comment);
 	    maclen = (4 + namelen +
 		      4 + enclen +
 		      4 + commlen +
@@ -1055,7 +1055,7 @@ unsigned char *openssh_loadpub(FILE *fp, char **algorithm,
         comment = dupstr(comment);
     }
 
-    pubblobsize = strlen(base64) / 4 * 3;
+    pubblobsize = (int)strlen(base64) / 4 * 3;
     pubblob = snewn(pubblobsize, unsigned char);
     pubbloblen = 0;
 
@@ -1074,7 +1074,7 @@ unsigned char *openssh_loadpub(FILE *fp, char **algorithm,
      * algorithm, and should match the encoded string at the start of
      * the public blob.
      */
-    alglen = strlen(line);
+    alglen = (int)strlen(line);
     if (pubbloblen < alglen + 4 ||
         GET_32BIT(pubblob) != alglen ||
         0 != memcmp(pubblob + 4, line, alglen)) {
@@ -1354,9 +1354,9 @@ int ssh2_save_userkey(const Filename *filename, struct ssh2_userkey *key,
 	unsigned char *macdata;
 	int maclen;
 	unsigned char *p;
-	int namelen = strlen(key->alg->name);
-	int enclen = strlen(cipherstr);
-	int commlen = strlen(key->comment);
+	int namelen = (int)strlen(key->alg->name);
+	int enclen = (int)strlen(cipherstr);
+	int commlen = (int)strlen(key->comment);
 	SHA_State s;
 	unsigned char mackey[20];
 	char header[] = "putty-private-key-file-mac-key";
@@ -1378,7 +1378,7 @@ int ssh2_save_userkey(const Filename *filename, struct ssh2_userkey *key,
 	putty_SHA_Init(&s);
 	putty_SHA_Bytes(&s, header, sizeof(header)-1);
 	if (passphrase)
-	    putty_SHA_Bytes(&s, passphrase, strlen(passphrase));
+			putty_SHA_Bytes(&s, passphrase, (int)strlen(passphrase));
 	putty_SHA_Final(&s, mackey);
 	hmac_sha1_simple(mackey, 20, macdata, maclen, priv_mac);
 	smemclr(macdata, maclen);
@@ -1391,7 +1391,7 @@ int ssh2_save_userkey(const Filename *filename, struct ssh2_userkey *key,
 	unsigned char key[40];
 	SHA_State s;
 
-	passlen = strlen(passphrase);
+	passlen = (int)strlen(passphrase);
 
 	putty_SHA_Init(&s);
 	putty_SHA_Bytes(&s, "\0\0\0\0", 4);
@@ -1487,7 +1487,7 @@ static char *ssh2_pubkey_openssh_str_internal(const char *comment,
 
     if (!alg) {
         alg = "INVALID-ALGORITHM";
-        alglen = strlen(alg);
+        alglen = (int)strlen(alg);
     }
 
     buffer = snewn(alglen +
@@ -1643,7 +1643,7 @@ static int key_type_fp(FILE *fp)
     int i;
     char *p;
 
-    i = fread(buf, 1, sizeof(buf)-1, fp);
+    i = (int)fread(buf, 1, sizeof(buf)-1, fp);
     rewind(fp);
 
     if (i < 0)
@@ -1668,7 +1668,7 @@ static int key_type_fp(FILE *fp)
         (p = p+1 + strspn(p+1, "0123456789"), *p == ' ') &&
         (p = p+1 + strspn(p+1, "0123456789"), *p == ' ' || *p == '\n' || !*p))
 	return SSH_KEYTYPE_SSH1_PUBLIC;
-    if ((p = buf + strcspn(buf, " "), find_pubkey_alg_len(p-buf, buf)) &&
+		if ((p = buf + strcspn(buf, " "), find_pubkey_alg_len((int)(p-buf), buf)) &&
         (p = p+1 + strspn(p+1, "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghij"
                           "klmnopqrstuvwxyz+/="),
          *p == ' ' || *p == '\n' || !*p))
