+++ src/filezilla/FtpControlSocket.h
@@ -36,7 +36,8 @@ CUSTOM_MEM_ALLOCATION_IMPL
     transfersize(0), transferleft(0),
     localFileHandle(INVALID_HANDLE_VALUE),
     bResume(FALSE), bResumeAppend(FALSE), bType(FALSE)
-  {}
+  {
+  }
   __int64 transfersize, transferleft;
   HANDLE localFileHandle;
   BOOL bResume, bResumeAppend, bType;
@@ -211,7 +212,7 @@ protected:
   bool m_skipReply;
 
   char * m_sendBuffer;
-  int m_sendBufferLen;
+  size_t m_sendBufferLen;
 
   bool m_bProtP;
